invoice.scanner.processing/requirements.txt
# ===== CORE DEPENDENCIES =====
# Task queue and message broker
celery>=5.3.0
redis>=5.0.0

# Web framework for HTTP service
flask>=3.0.0

# Database
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0

# ===== IMAGE PROCESSING =====
opencv-python-headless<=4.6.0.66
Pillow>=10.0.0
numpy>=1.26.2
pdf2image>=1.16.0
PyPDF2>=3.0.0

# ===== OCR - TEXT EXTRACTION =====
# Tesseract (CPU-based OCR)
pytesseract>=0.3.10

# PaddleOCR (CPU/GPU capable, better accuracy)
# GPU support added in Dockerfile.ml
paddlepaddle>=3.0.0
paddleocr>=2.7.0

# ===== LLM PROVIDERS =====
# OpenAI
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
openai>=0.27.0

# Google Gemini
langchain-google-genai>=0.0.1
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.0.0

# Anthropic Claude
langchain-anthropic>=0.1.0
anthropic>=0.7.0

# ===== DATA PROCESSING & VALIDATION =====
pydantic>=2.7.0
marshmallow>=3.20.0
pandas>=2.0.0

# ===== LOGGING & MONITORING =====
python-dotenv>=1.0.0
structlog>=23.0.0
loguru>=0.7.0

# ===== UTILITIES =====
requests>=2.31.0
python-dateutil>=2.8.0
